feat: apply state management to the Linear Progress Bar using useState and useEffect
