+++ src/components/ProgressBar/progress_bar.module.scss
@@ -1,5 +1,5 @@
 .sect__container {
-    width: 80%;
+    width: 90%;
     max-width: 1024px;
     margin: 0 auto;
 }
@@ -22,9 +22,9 @@
         // padding: .3em 0;
     }
 
-    .progress__percent {
-        min-width: 20%;
-    }
+    // .progress__percent {
+    //     min-width: 20%;
+    // }
 
     .percent__number {
         color: #6B7280;
