+++ src/pages/Home/index.js
@@ -3,7 +3,7 @@ import NavBar from "../../components/NavBar";
 import UploadBox from "./components/UploadBox";
 import FileLoaded from "./components/FileLoaded";
 import LoadingFile from "./components/LoadingFile";
-import ProgressBar from "../../components/ProgressBar";
+import ProgressBar from "../../components/ProgressBar/Index";
 
 const Home = () => {
   return (
